refactor(gemini): move session manager into provider package
